Function Documentation

DataAssimilation.EKFsolve — Method

EKFsolve(prob::AssimilationProblem)

Given an AssimilationProblem prob, obtain the analysis uₐ and error covariance matrix B using the Extended Kalman Filter (EKF) method.

DataAssimilation.model_forward! — Method

model_forward!(uₖ, uₐ, t_now, t_next, prob)

Given value uₖ and time range tnow and tnext, integrate the ODE problem forward and update solution object.

DataAssimilation.timeindex — Method

timeindex(t, Δt)

For a given time value and time step, return the corresponding array index.

DataAssimilation.update! — Method

update!(sol, t_now, t_next, Δt, uₐ)

Update solution array uₐ with the values from the solution object sol at specified time step Δt